GamePad Communication Code - Linux Python 3.6 EvDev
**IMPORTANT
The following also needs to be installed in your Linux platform: 
Python 3.6
EvDev

This line in the code:
game_pad = InputDevice('/dev/input/event12')
must be modified when working with a new computer. The 'event#' part is the gamepad itself. First, unplug the gamepad from the computer and run the following command in the terminal: python -m evdev.evtest . Make a note of which event number show up, and then plug in the gamepad and run the command again. The new event that shows up is the event associated with the gamepad. Replace that event# in the line of code mentioned.